feat(theme): sync window chrome with QueryaThemeScope workbench tokens
Drive title bar background and bitsdojo window button colors from workbench
canvas, surface, and mutedForeground instead of shadcn ColorScheme alone.
